push pertemuan 4 - update

+++ Pertemuan 4/Praktek/src/com/polban/jtk/sales/Product.java
@@ -16,7 +16,7 @@ public class Product {
     }
 
     public double getPrice() {
-        return price;
+        return (long) price;
     }
 
     public void setPrice(double price) {
